setup: Don't assume exe is built on --lib
Cabal-3.12 introduces a breaking change which broke `cabal install --lib Agda`.

The reason is that previously, Cabal, for a Custom-build-type package,
would /always/ build the executable, even if only the library was
requested, and Agda's Setup.hs copyHook expects the executable to
/always/ be available to compile the Agda standard lib. Now, Cabal won't
build the executable if it is not required, so Agda's assumption will
cause a crash when the executable isn't built.

The fix is simple: don't try to compile the Agda standard library if
only the Agda Haskell library was requested, since only when installing
the executable does the agda standard library need to be installed.

Co-authored-by: @sheaf
Co-authored-by: @andreasabel