Make BackendName a Text rather than a String
Main motivation for this refactoring was to consistently enforce the
type synonym `BackendName`; so I changed it from `String` to `Text`.

In the long run, we should deprecate `String` for `Text` everywhere,
but this will be another mammut refactoring.

Also changed the "installed backends" list into a bullet list in the
error for non-existing backend.