Make termination checking more permissive wrt non-exact clause reduction
When the termination checker is trying to reduce away function calls,
block the reduction of recursive clauses when they are successfully
matched instead of removing both recursive and non-exact clauses from
consideration altogether. This way, a non-exact clause can still be
used to reduce away a function call as long as the clauses before it
don't match.