Refactor: move defArgGeneralizable to GeneralizableVar
This field, pertaining to any Definition, is actually only used for generalizable variables.
Thus, placing it just there is more economical and clarifies its purpose.

Originally, it might have been intended for other definitions as well, but the implementation of generalization has been stable for a good while now, so we can tighten the design.