Fix compilation of serialisation code on 32 bit platforms
The "Hashable Node" implementation uses a 64-bit hash. However, this
does not work on 32 bit platforms (as the 64-bit constants do not fit
into a 32 bit word).

This commit uses alternative constants when on 32-bit platforms. The
hash factor is the 32-bit fibonacci constant, and the salt is
arbitrarily chosen.