Fix compilation of serialisation code on 32 bit platforms (#7559)
* Fix compilation of serialisation code on 32 bit platforms

The "Hashable Node" implementation uses a 64-bit hash. However, this
does not work on 32 bit platforms (as the 64-bit constants do not fit
into a 32 bit word).

This commit uses alternative constants when on 32-bit platforms. The
hash factor is the 32-bit fibonacci constant, and the salt is
arbitrarily chosen.

* Explain the source of some of the constants

The salt really does just seem to be "I made it up". It's copied and
pasted several places, but just goes back to Andras!