Refactor: use Set1 instead of DList in UsedNames and ShadowingNames
In the end, we always want a Set there, so why not use one from the
start?

The current "optimization" via DList runs the risk that this DList is
converted to a list many times, so ruining any benefit over using a
Seq or Set.

NB: the sets here should be small in practice, so maybe all of this
does not matter, and it could have remained a plain list which is what
it was in the beginning.  But a DList does not feel right here...