Refactor: SourceFile is FileId instead of AbsolutePath; add FileDict to TCM
We introduce a bimap from AbsolutePath to FileId into the persistent TCState.
This map is supposed to know about all the files we read during an
Agda session, and translate their names to unique IDs.

This extra structure is supposed to help getting rid of storing and
passing around a lot of AbsolutePath.
Later the IDs shall also identify top-level modules.