Refactor: cache info about prim. modules; isBuiltinModule FileId
When we translate an AbsolutePath to a FileId, we also check whether
the path belongs to one of Agda's builtin modules, and remember this
information in the SessionTCState.

This allows for quick and uncomplicated checking of whether we are in
a builtin module while scope checking.

This should remove the number of file system accesses done by Agda.