Setup: unconditionally check if we want interfaces
Previously we only checked if the Cabal library was new enough to trigger the bug, but this bug is triggered by the version of cabal-install (the build tool) not Cabal (the library). These two can come apart, and it's valid to use an older version of Cabal. So instead we unconditionally check if interfaces are wanted.

Resolves https://github.com/agda/agda/issues/7655

See also https://github.com/agda/agda/pull/7471