Block instances on discrimination tree overlap
Fixes a performance regression introduced by #7298 to the 1Lab. The
issue boils down to a constraint like

    _0 : Extensional ∣ ₀ _P_669 y ∣ _ℓr_724

where the bars and ₀ are both projections. In the module which prompted
this fix, there are *29* total 'Extensional' instances, and, of course,
they would all be tried for this constraint. Moreover, since none of the
candidates can be rejected until _P_669 is solved, this constraint would
be retried over and over, resulting in

    % grep 'checkCandidateForMeta _725@' [debug output] | wc -l
    260

quite a few useless retries! With this patch, we wait until _P makes
progress, so only *one* candidate is checked for this meta.

In this module, the time spent in checkCandidateForMeta is cut by 50%.