fixup: reduce head instance before dropping dupes
Fixes an issue where the standard library subtly relies on the fact that
instance candidates get reduced when an instance constraint gets
postponed to implement overlapping behaviour. More specifically, they
have

    coerce-maybe : ⦃ Coercible A B ⦄ → Coercible (Maybe A) (Maybe B)
    coerce-maybe _ = TrustMe

    coerce-refl : Coerce A A
    coerce-refl = TrustMe

and for the code to type-check we *have* to reduce coerce-maybe to a
lambda, otherwise we don't have TrustMe ≠ TrustMe to "resolve" the
overlap.