Test cases for #7486
Treats the new behaviour (instances for a reducible function are
wildcardsa) as correct, and adds the previously-accepted, now-failing
programs that would silently ignore all the possible instances as
failing test cases.

Adds a successful test case for adding a pair of instances at once by
branching on a boolean in the return type.