Re #7732: use `agda --emacs-mode` in agda2-set-program version
Only for switching to agda older than 2.8 do we still invoke `agda-mode locate`,
for other ones we invoke `agda --emacs-mode locate` instead.

Note that switching from 2.7 to 2.8 will still invoke `agda-mode`
since the code for the switcher is the one of 2.7.
Thus, it is good to keep around the `agda-mode` executable for a
while, even in future releases of Agda.