re #3209: print out-of-scope names unqualified in case splits
Names coming from parameterised modules might not be found by the
abstract-to-concrete lookup, even when they are in scope (see issue #3209).
This leads to poor user experience when splitting on data types
defined in parameterised modules.

While fixing the underlying issue would probably require broad changes
to the module system, we can at least make user experience a bit better
by stripping the module prefix when generating clauses for a case split,
in which case the constructors are more likely to be in scope
unqualified than fully qualified.