Re #7587: removed broken AbsurdLam heuristics from Mimer
The current heuristics only worked in the case `\bot -> A` and
produced garbage for all other cases where such type arises as a
subgoal during proof search.

The current implementation was not close to a working one, so I
consider this a non-solution which gets a "reject" rather than "accept
with revisions".

Closes #7587 symptomatically, but a proper fix would be nice to see.