Fix #7832 by checking for eta record constructor in properlyMatching
This places `properlyMatching` in a `HasConstInfo` monad.
Consequently, the computation of occurrences in the positivity checker
has to be placed in such a monad.
Here, we use `TCM`, but one could use something more lightweight, like
`ReduceM` or a custom monad that just supports `HasConstInfo` and nothing
more.
TODO: benchmark the positivity checker, and see whether making it
strict (in `TCM`) has some adversary effects on performance.

Closes #7832.