Build and test WASM in CI
This is a little nasty, as WASM/WASI cannot spawn subprocesses. As the
test suite largely involves around spawning new Agda instances, we
instead run the native test suite, but using the Agda WASM module as the
binary.

This unfortunately requires compiling Agda twice (once for WASM, once
for the native platform).

 - Add a flag (or rather, environment variable) to skip tests that spawn
   subprocesses.

 - Add a new wasm.yml workflow, that uses the Nix flake to compile and
   test WASM.