Refactor absTerm to arbitrary de Bruijn index for hole, was just 0
The implementation of absTerm always freed de Bruijn index 0 for the
placeholder of the abstracted-out subterm, but then needed to `swap01`
swap de Bruijn indices 0 and 1 just before putting back lambdas (and
other binders).

It seems more principled to directly implement `absTerm j u` that
replaces subterm `u` with de Bruijn index `j`, shifting indices up to
make space for `j`.  When going under a binder, we simply increase `j`.

Now `absTerm j u` is the inverse of single substitution `subst j u`.

The new implementation seems marginally faster than the old
implementation, but measurement is difficult because of the overall
time for `with`-abstraction is dominated by the time to check that the
generated type of the `with`-function is well-formed (issue #7975).