Improve module checkpointing performance (#7951)
* [perf] add UNPACK pragma to NameId in Abstract.Name

In my benchmarking, this shaves about 11 seconds off of a clean build
of the 1lab.

* [debug] Add verbose print for difference between module checkpoints

* [perf] Replace module checkpoint map with a stack

This reduces runtime by about 5% in my benchmarks, as we get
to avoid doing repeated calls to `Map.union` inside of `checkpoint`.

* [fix] Fix wayward comment

* [fix] dont unwind past the checkpoint id we are looking for

* [fix] Fix ordering in Pretty instance of ModuleCheckpoints

* [fix] More precise module stack unwinding code

* [fix] Fix checkpoint unwinding after TCState resets

* [test] Add test for interleaving modules inside records and fields

This test is a minimal reproducer of the structure of `Cat.Bi.Base`
from the 1lab, which fails to typecheck with the optimized module
checkpointing code.