WIP: fix #8033: color warnings by keeping Doc over rendering to String
Here I change `MonadDebug` to keep `Doc` rather than rending it to
`String`, so that it can be stored in `Resp_RunningInfo` and the
client can decide how to render it.
For now, the console renderer `putDoc` turns the annotations in `Doc`
to highlighting, while nothing is implemented for Emacs.

The changes here seem to introduce a massive space leak, running the
testsuite that spawns several Agda processes gets my terminal to
freeze.

So I do not expect CI to be able to run the tests at this stage.

I have no clue where the space leak comes from, help is appreciated.

TODOs:
- [ ] fix space leak
- [ ] spurious newlines are introduced, e.g. after the warning in `test/Fail/Issue4723.err`