Re #659: link from warnings to documentation (if your terminal supports it)
This equips the name of the warning with a link to agda.readthedocs.io.
Not all terminals support hyperlinks (OSC 8) though, e.g. macOS standard Terminal does not, though iTerm does.

Printing the URL so it is usable on Terminal etc. seems too noisy.

Disclaimer: these links are only correct for the released version of Agda for now, see PR #8032.