--color=never disables info buffer highlighting, keep from golden values
If the text for the Agda information buffer contains filenames, which
are absolute pathes, the highlighting info is not portable (e.g. from
local runs to CI).

Thus, we utilize option @--color=never@ to turn off highlighting in
the information buffer.
We turn the option on in the interaction tests to keep this highlighting
info out of golden values.