Fix #8048 by generating Lock constraints also for implicitArgs
This commit factors out the lock constraint generation and generalizes
`namedImplicitArgs` to `checkedImplicitArgs` including possible lock
constraints for the inserted arguments.

`namedImplicitArgs` now makes sure no constraints are generated and
crashes with a `NotSupported` error if that is not the case.

There is probably opportunity for handling the constraints in other
places where we use `implicitArgs`, but this remains TODO if there is
public demand for it.

For now, we at least addressed the unsoundness, and constraints are
supported for implicit arguments inserted during `checkApplication`.

Closes #8048.