Refactor: move InteractionOutputCallback to SessionState
The `InteractionOutputCallback` value is never changed during a run of
Agda, it is set initially by the interactor which cannot be changed
during an Agda session.

Thus, we move it to the `TCSessionState` so that we need not restore
it after resetting the state (e.g. when a different module is
loaded into Agda).