clean forget-all-goals
This fixes several problems with agda2-forget-all-goals.

The main problem was that it was removing all display properties, which needlessly break many minor modes. Now it removes only the 'category properties, which are the only ones that agda mode creates anyway.

Besides, the previous code was doing many useless things:
1. removing properties which don't exist
2. using annotation-preserve-mod-p-and-undo, even though remove-text-properties does not change the undo data
3. finding the overlays by testing property changes, which is error-prone.

So I cleaned those at the same time.