[refactor] generalize freezeMetas to Traversable MetaId collection
The use of `LocalMetaStore` both for the `stOpenMetaStore` and the
collection `ms` of metas to freeze confused me, so I generalized the
function to take any collection `ms` of `MetaId`s.

Being more robust, it is also more applicable, as we do not need to
pass a `LocalMetaStore` but just the ids of the metas we want to
freeze.  From that store we anyway just used the `keys` list.

(It just happened that these metas were always organized as
`LocalMetaStore` when we wanted to call `freezeMetas`,
so may that explains that we made `freezeMetas` take `LocalMetaStore`.
But it passes in **too much** information and is thus not a good fit.)