Fix #8103 by removing the parsing hack in BasicOps.metaHelperType
The internal error of issue #8103 exposed the fragility of the parsing heuristics for the helper function template via an ugly prefixing with `let f = _ in`.

We replace it by the natural compositional approach:
1. Parse as concrete syntax.
2. Operator-parse into simple concrete application `f e1 ... en`.
3. Scope and type check the arguments `e1...en` but keep the new name `f` as just a `C.Name`.

We make the type of `metaHelperType` more precise, now constructing an `OutputConstraint' A.Expr C.Name` rather than an `OutputConstraint' A.Expr A.Expr`.
This also refines the type of the constructor `Goal_HelperFunction` which might be breaking 3rd-party interaction frontends.

Closes #8103.