Fix #7646: use 'Range' from interaction meta if goal command supplies no range
This affects the golden values in the interaction testsuite in many
places, but there are only two kind of changes:

- a correct range when there was no range or a bogus range without filename
- more precise `Give_NoParen` (disabled when no range is given) instead of `Give_String`

Actual interaction in an editor should not be affected since it always
supplies a range with a goal command; only the testsuite provides no
ranges.

In effect, we make the test-suite more realistic since responses are
now akin the responses sent to the editor.