Faster convergence in case of termination errors
Instead of first completing the call graph and then checking for bad
loops, we now check for bad loops in all newly added calls during
completion.  Thus, as soon as we hit a counterexample to the
termination condition, we can stop graph completion.

This might not be so relevant in practice since in almost all code
termination checkings succeeds and thus has to do the whole
completion.

But I can think of scenarios where the user raises the
termination-depth in hope of getting termination through and might
have to wait longer to get the result (that termination checking still
fails).

The other advantage is that one might get smaller counter-examples, as
e.g. in Issue1075 (updated golden value enclosed).