Generate parts of the Agda input method procedurally
Instead of having to write out all input mappings, this change
introduces functions to lookup and automatically generate charter
mappings of a certain form based on Emacs' knowledge of unicode
character names.  So

  (agda-input-latin-range "Mi%s" "MATHEMATICAL ITALIC CAPITAL %s")

replaces the previous listing of

  ("MiA" "𝐴")
  ("MiB" "𝐵")
  ("MiC" "𝐶")
  ("MiD" "𝐷")
  ("MiE" "𝐸")
  ...

This change also cleans up some related documentation and replaces
some functional-style definitions with imperative ones that should put
less pressure on GC and on the stack.  It is therefore possible to
remove the risky rebinding of `max-lisp-eval-depth`.