Merge annotation.el into agda2-highlight.el
The motivation for dropping annotation is that it replicates
functionality that Xref provides by default.  By replacing the custom
goto-definition system with the standard Xref interface, the need for
a library like annotation goes away.

Note that this is not a complete change, and it has placeholders from
the places where the annotate.el functionality will be replaced by
Xref.

This patch also addresses minor issues in agda2-highlight.