Refactor: run termination analysis with dot-patterns from the start
This commit removes the optimization that we first extract calls
without considering dot-patterns, and only if the termination check
fails we run everything again, taking dot-patterns into account this
time.

The savings of this optimization are negligible, and we can get a
simpler structure of the termination checker if we remove it.

Standard-library test with dot-pattern optimization:

Total                                   193,958ms
Termination                               1,022ms  (3,874ms)
Termination.RecCheck                      2,342ms
Termination.Compare                         306ms
Termination.Graph                           203ms

After removing the optimization:

Total                                   194,746ms
Termination                               1,061ms  (3,928ms)
Termination.RecCheck                      2,319ms
Termination.Compare                         368ms
Termination.Graph                           179ms

It seems that we spend a bit more time in Termination.Compare (since
we also have to relate call arguments to the dot patterns of the
caller), but it is not even a tenth of a second.