Undo "Faster convergence for high --termination-depth"
That heuristics loses termination certification for the somewhat
contrived example I added to
`test/Succeed/TerminationWithTwoConstructors.agda`,
so we disable it.

As a consequence, we have to be careful with high
`--termination-depth` again for cases that do not termination
check (because even iterative deepening might take long to finish if
in the end counterexamples remain).