UselessImport warning instead of ParseError for `import M x y`
The desugaring of `import` with direct instantiation of the module
parameters into a plain `import` and a module macro has been moved
from the parser to the nicifier.

Instead of a hard parse error we can now issue a warning and ignore
the useless import statement.

The desugaring should be further moved into the scope checker so that
we can get rid of the hack of generating fresh concrete module names
using hashes.